Refactor packages in project
